requirements.txt
Flask==2.2.3
pyaudio==0.2.11
pydub==0.25.1
Pillow==9.3.0
Werkzeug==2.2.3
opencv-python==4.7.0.72
numpy==1.24.2
pandas==1.5.3
matplotlib==3.7.0
torch==1.13.1
torchvision==0.14.1
random==1.0
pygame==2.1.2
base64==1.0.0